refactor: better colored feature flag
