+++ .github/workflows/ci.yml
@@ -81,7 +81,7 @@ jobs:
           cargo run --example simple_stdout
           cargo run --example multiple_dispatches
           cargo run --example custom_layout_filter
-          cargo run --features="no-color" --example simple_stdout
+          cargo run --no-default-features --example simple_stdout
           cargo run --features="json" --example json_stdout
           cargo run --features="json,rolling-file" --example rolling_file
           cargo run --features="fastrace/enable" --example fastrace
