+++ Cargo.toml
@@ -32,11 +32,13 @@ all-features = true
 rustdoc-args = ["--cfg", "docsrs"]
 
 [features]
+default = ["colored"]
+
+colored = ["dep:colored"]
 fastrace = ["dep:fastrace"]
 journald = ["dep:libc"]
 json = ["dep:serde_json", "dep:serde", "jiff/serde"]
 native-tls = ["dep:native-tls", "fasyslog?/native-tls"]
-no-color = ["colored/no-color"]
 non-blocking = ["dep:crossbeam-channel"]
 opentelemetry = [
   "dep:opentelemetry",
@@ -48,12 +50,12 @@ syslog = ["non-blocking", "dep:fasyslog"]
 
 [dependencies]
 anyhow = { version = "1.0" }
-colored = { version = "3.0" }
 env_filter = { version = "0.1.1" }
 jiff = { version = "0.1.14" }
 log = { version = "0.4", features = ["std", "kv"] }
 
 # Optional dependencies
+colored = { version = "3.0", optional = true }
 crossbeam-channel = { version = "0.5", optional = true }
 fastrace = { version = "0.7", optional = true }
 fasyslog = { version = "0.3", optional = true }
