+++ src/diagnostic/mod.rs
@@ -12,7 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-//! Mapped Diagnostic Context (MDC). A lighter technique consists of uniquely stamping each log request.
+//! Mapped Diagnostic Context (MDC). A lighter technique consists of uniquely stamping each log
+//! request.
 
 use std::borrow::Cow;
